simulation/requirements.txt
decorator==4.4.0
networkx==2.3
netaddr==0.7.19
semver==2.9.0
cachetools==4.0.0
multipledispatch==0.6.0
py-flags==1.1.4

toml==0.10.2
rtoml==0.7.0
setuptools_rust==1.1.2
pyyaml==5.4.1
gura==1.2.0

Jinja2~=3.0.1
stringcase~=1.2.0
typing-inspect==0.7.1

redis==3.5.3
rejson==0.5.4